+++ app/Http/Controllers/Superadmin/UserController.php
@@ -6,6 +6,10 @@ use App\Http\Controllers\Controller;
 use Illuminate\Support\Facades\Hash;
 use App\Models\User;
 use App\Models\Role;
+use App\Models\Order;
+use App\Models\Design;
+use App\Models\Spk;
+use App\Models\Production;
 use Illuminate\Http\Request;
 
 class UserController extends Controller
@@ -39,6 +43,24 @@ class UserController extends Controller
 
     public function store(Request $request)
     {
+        $existsUser = User::where('username', $request->username)->exists();
+
+        if ($existsUser) {
+            return response()->json([
+                'success' => false,
+                'message' => 'Username already in use!'
+            ], 400);
+        }
+
+        $existsEmail = User::where('email', $request->email)->exists();
+
+        if ($existsEmail) {
+            return response()->json([
+                'success' => false,
+                'message' => 'Email already in use!'
+            ], 400);
+        }
+
         $request->validate([
             'role_id'   => 'required|exists:roles,id',
             'full_name' => 'required|string|max:100',
@@ -73,43 +95,50 @@ class UserController extends Controller
 
     public function update(Request $request, $userId)
     {
-        $user = User::find($userId);
+        $existsUser = User::where('username', $request->username)
+        ->where('id', '!=', $userId)
+        ->exists();
 
-        if (!$user) {
+        $existsEmail = User::where('email', $request->email)
+        ->where('id', '!=', $userId)
+        ->exists();
+
+        if ($existsUser) {
             return response()->json([
                 'success' => false,
-                'message' => 'User not found'
-            ], 404);
+                'message' => 'Username already in use!'
+            ], 400);
+        }
+
+        if ($existsEmail) {
+            return response()->json([
+                'success' => false,
+                'message' => 'Email already in use!'
+            ], 400);
         }
 
         $request->validate([
-            'role_id'   => 'nullable|exists:roles,id',
-            'full_name' => 'nullable|string|max:100',
-            'username'  => 'nullable|string|max:50|unique:users,username,' . $userId,
-            'email'     => 'nullable|email|unique:users,email,' . $userId,
-            'user_status' => 'nullable|in:Pending,Active,Non-Active'
+            'role_id'   => '|exists:roles,id',
+            'full_name' => 'required|string|max:100',
+            'username'  => 'required|string|max:50|unique:users,username,' . $userId,
+            'email'     => 'required|email|unique:users,email,' . $userId,
+            'user_status' => 'required|in:Pending,Active,Non-Active'
         ]);
 
-        $update = [];
+        $user = User::findOrFail($userId);
 
-        if ($request->has('role_id')) $update['role_id'] = $request->role_id;
-        if ($request->has('full_name')) $update['full_name'] = $request->full_name;
-        if ($request->has('username'))  $update['username'] = $request->username;
-        if ($request->has('email'))     $update['email'] = $request->email;
-        if ($request->has('user_status')) $update['user_status'] = $request->user_status;
-
-        $user->update($update);
+        $user->update($request->all());
 
         return response()->json([
             'success' => true,
-            'message' => 'User updated',
+            'message' => 'User updated successfully',
             'data' => $user
         ]);
     }
 
     public function resetPassword($userId)
     {
-        $user = User::find($userId);
+        $user = User::findorFail($userId);
 
         if (!$user) {
             return response()->json([
@@ -134,13 +163,31 @@ class UserController extends Controller
 
     public function destroy($userId)
     {
-        $user = User::find($userId);
+        $user = User::findorFail($userId);
 
         if (!$user) {
             return response()->json([
                 'success' => false,
                 'message' => 'User not found'
             ], 404);
+        }
+
+        if ($user->user_status == 'active') {
+            return response()->json([
+                'success' => false, 
+                'message' => 'Cannot delete active user!'], 400);
+        }
+
+        if (
+            Order::where('created_by', $user->id)->exists() ||
+            Design::where('assigned_to', $user->id)->exists() ||
+            Spk::where('assigned_to', $user->id)->exists() ||
+            Production::where('assigned_to', $user->id)->exists() 
+        ) {
+            return response()->json([
+                'success' => false,
+                'message' => 'User is used in other data!'
+            ], 400);
         }
 
         $user->delete();
